Added new tables to the database.
The tabes which relate to the given user are used as alternatives for authentication. Oauth and local accounts will both be possible avenues for account creation.

JSOn Web Tokens will be used to authorize the client and the contents of the web token shuld include the ID of the authorized user.