Database Refactor (#9)
* Added new tables to the database.

The tabes which relate to the given user are used as alternatives for authentication. Oauth and local accounts will both be possible avenues for account creation.

JSOn Web Tokens will be used to authorize the client and the contents of the web token shuld include the ID of the authorized user.

* Confirmed that the Google auth still functions as intended.

Removed player schema as we will be using JWT for route authorization.

Updated README's Database and TODO sections.

* Included issue template for bug reporting.

* Fixed typo.

* Added feature request template.

* Implement auth API (register, login) and user API routes (#10)

* Implement auth API and merge database changes

* Adjusted the auth route so it uses the updated database schema.

Added SALT_ROUNDS variable to adjust hashing parameters from the environment.
Defaulted database location to the one defined in env.

Updated README.

* Database no longer tracks the avatar's URL provided by Oauth providers.

It's possible to add them to the provider table or have  alocal one if we would like to display user avatars.

* Updated API.md based on the refactored database.

* Merge branch 'database' into Standard-user-management

* Removed credential inclusion from user in the login route as it exposed the pasword to the originator when sending back a successful reply.