Refactored the Game component in Backend (#30)
* Added a route for retrieving player statistics.

Updating the stats will have to happen in the backend game logic.

* Added linting and formatting to the backend.

This is similarly based on a personal clang-format adapted from a heavily modified Google style guideline.

* Checking if username is already taken during registration

* Included protected route for looking up specific player stats without exposing additional details.

* Resolved the issue of a parameter name mismatch.

* Applied formatting to the seed file.

* Added utility functions to verify password strength and email format.

* Added extension recommendations back in

* Integrated the pong microservice into the backend server code.

Removed excessive Dockerfile and modified the docker-compose file.
Modified makefile and removed the game container target.

Adjusted listening path for the game on the frontend.

Converted game-related javasccript files to typescript and made them conform to ECMAScript Modules.

Registered game logic within the core of the backend.